Add new sdv.relational subpackage (#198)
