+++ sdv/errors.py
@@ -1,0 +1,5 @@
+"""SDV Exceptions."""
+
+
+class NotFittedError(Exception):
+    """Error to raise when sample is called and the model is not fitted."""
