+++ sdv/relational/__init__.py
@@ -1,0 +1,5 @@
+"""Relational models subpackage."""
+
+from sdv.relational.hma import HMA1
+
+__all__ = ['HMA1', ]
